+++ src/Components/Bucket/BucketCard.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import { Card, Col, Modal, Row, Statistic, Tag } from "antd";
 import humanizeDuration from "humanize-duration";
 import { Bucket, BucketInfo, Client, TokenPermissions } from "reduct-js";
@@ -8,8 +8,8 @@ import prettierBytes from "prettier-bytes";
 import { DeleteOutlined, SettingOutlined } from "@ant-design/icons";
 
 import "./BucketCard.css";
-import CreateOrUpdate from "./CreateOrUpdate";
-import RemoveConfirmationByName from "../RemoveConfirmationByName";
+import BucketSettingsForm from "./BucketSettingsForm";
+import RemoveConfirmationModal from "../RemoveConfirmationModal";
 import { bigintToNumber } from "../../Helpers/NumberUtils";
 
 interface Props {
@@ -35,9 +35,14 @@ export const getHistory = (interval: {
 export default function BucketCard(props: Readonly<Props>) {
   const [confirmRemove, setConfirmRemove] = useState(false);
   const [changeSettings, setChangeSettings] = useState(false);
-  const { client, bucketInfo, index } = props;
+  const [bucketInfo, setBucketInfo] = useState(props.bucketInfo);
+  const { client, index } = props;
 
-  const onRemoved = async () => {
+  useEffect(() => {
+    setBucketInfo(props.bucketInfo);
+  }, [props.bucketInfo]);
+
+  const onRemove = async () => {
     const bucket: Bucket = await client.getBucket(bucketInfo.name);
     await bucket.remove();
     props.onRemoved(bucketInfo.name);
@@ -104,10 +109,10 @@ export default function BucketCard(props: Readonly<Props>) {
           />
         </Col>
       </Row>
-      <RemoveConfirmationByName
+      <RemoveConfirmationModal
         name={bucketInfo.name}
-        onRemoved={onRemoved}
-        onCanceled={() => setConfirmRemove(false)}
+        onRemove={onRemove}
+        onCancel={() => setConfirmRemove(false)}
         confirm={confirmRemove}
         resourceType="bucket"
       />
@@ -117,7 +122,7 @@ export default function BucketCard(props: Readonly<Props>) {
         footer={null}
         onCancel={() => setChangeSettings(false)}
       >
-        <CreateOrUpdate
+        <BucketSettingsForm
           client={client}
           key={bucketInfo.name}
           bucketName={bucketInfo.name}