+++ src/Components/RemoveConfirmationModal.tsx
@@ -1,43 +1,49 @@
 import React, { useEffect, useState } from "react";
 import { Form, Input, Modal } from "antd";
 
-interface Props {
+interface RemoveConfirmationModalProps {
   name: string;
-  onRemoved: () => void;
-  onCanceled: () => void;
+  onRemove: () => void;
+  onCancel: () => void;
   resourceType: string;
   confirm: boolean;
 }
 
-export default function RemoveConfirmationByName(props: Readonly<Props>) {
+export default function RemoveConfirmationModal({
+  name,
+  onRemove,
+  onCancel,
+  resourceType,
+  confirm,
+}: RemoveConfirmationModalProps) {
   const [confirmRemove, setConfirmRemove] = useState(false);
   const [confirmName, setConfirmName] = useState(false);
 
   useEffect(() => {
-    setConfirmRemove(props.confirm);
-  }, [props.confirm]);
+    setConfirmRemove(confirm);
+  }, [confirm]);
 
   const checkName = (bucketName: string) => {
-    setConfirmName(bucketName == props.name);
+    setConfirmName(bucketName == name);
   };
 
   return (
     <Modal
       open={confirmRemove}
       onOk={() => {
-        props.onRemoved();
+        onRemove();
         setConfirmRemove(false);
       }}
-      onCancel={props.onCanceled}
+      onCancel={onCancel}
       closable={false}
-      title={`Remove ${props.resourceType} "${props.name}"?`}
+      title={`Remove ${resourceType} "${name}"?`}
       okText="Remove"
       confirmLoading={!confirmName}
       okType="danger"
       data-testid="delete-modal"
     >
       <p>
-        For confirmation type <b>{props.name}</b>
+        For confirmation type <b>{name}</b>
       </p>
       <Form.Item name="confirm">
         <Input