+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- RS-509: add option to rename a bucket in the console, [PR-70](https://github.com/reductstore/web-console/pull/70)
+
 ## [1.7.0] - 2024-08-19
 
 ### Added: